@@ -6,12 +6,12 @@ public class ProfileDataHandler : IProfileDataHandler
     public bool IsViaTeacher(string userName)
     {
         if (string.IsNullOrEmpty(userName)) return false;
-        if (IsNotVia(userName)) return false;
+        if (!IsViaAccount(userName)) return false;
         
         return true;
     }
 
-    private static bool IsNotVia(string userName)
+    private static bool IsViaAccount(string userName)
     {
         return userName.StartsWith("VIA\\");
     }
