@@ -1,5 +1,5 @@
 ﻿using Application.Features.CreateCategory;
-using Application.Features.ProfileDataLogic.DTOs.CategoryDTOs;
+using Application.Features.SharedDtos;
 using SharedKernel.Results;
 
 namespace Application.Features.OldStuffToBeMovedAround.LogicContracts;
