@@ -5,7 +5,7 @@ namespace Application.RepositoryContracts;
 public interface ICategoryRepo
 {
     Task<Category> AddToTeacher(Category category, string teacherName);
-    Task<ICollection<Category>> GetCategoriesByTeacherAsync(string teacherId);
+    Task<IEnumerable<Category>> GetCategoriesByTeacherAsync(string teacherId);
     Task UpdateAsync(Category categoryToUpdate);
     Task DeleteAsync(Guid categoryId);
     Task<Category> GetCategoryById(Guid id);
