@@ -1,5 +1,4 @@
 ﻿@using BlazorServerUI.Pages.Profile.RadialMenuItems
-@using Application.Features.ProfileDataLogic.DTOs.CategoryDTOs
 @using Application.Features.ProfileDataLogic.DTOs.ExternalResourceDTOs
 @using Application.Features.ProfileDataLogic.DTOs.GuideDTOs
 @using Application.Features.ProfileDataLogic.LogicContracts
@@ -7,6 +6,7 @@
 @using BlazorServerUI.StateContainers.Profile
 @using Domain.Exceptions
 @using Application.Features.ProfileDataLogic.DTOs
+@using Application.Features.SharedDtos
 @using BlazorServerUI.Popups
 
 @inject IGuideProvider guideProvider
