@@ -1,8 +1,8 @@
-﻿@using Application.Features.ProfileDataLogic.DTOs.CategoryDTOs
-@using Application.Features.ProfileDataLogic.LogicContracts
+﻿@using Application.Features.ProfileDataLogic.LogicContracts
 @using BlazorServerUI.StateContainers.Profile
 @using JsonData.Repositories
 @using Application.Features.OldStuffToBeMovedAround.LogicContracts
+@using Application.Features.SharedDtos
 @using BlazorServerUI.Popups
 @namespace BlazorServerUI.Pages.Profile.RadialMenuItems
 @inject ICategoryLogic categoryLogic
