@@ -1,8 +1,8 @@
 ﻿@using BlazorServerUI.StateContainers.Profile
 @using System.Text.Json
-@using Application.Features.ProfileDataLogic.DTOs.CategoryDTOs
 @using Application.Features.ProfileDataLogic.DTOs.ExternalResourceDTOs
 @using Application.Features.ProfileDataLogic.LogicContracts
+@using Application.Features.SharedDtos
 @using BlazorServerUI.Popups
 
 @namespace BlazorServerUI.Pages.Profile.RadialMenuItems
