@@ -1,7 +1,7 @@
 ﻿@using BlazorServerUI.StateContainers.Profile
-@using Application.Features.ProfileDataLogic.DTOs.CategoryDTOs
 @using Application.Features.ProfileDataLogic.DTOs.GuideDTOs
 @using Application.Features.ProfileDataLogic.LogicContracts
+@using Application.Features.SharedDtos
 @using BlazorServerUI.Popups
 @namespace BlazorServerUI.Pages.Profile.RadialMenuItems
 @inject IGuideLogic guideLogic
