@@ -1,13 +1,22 @@
 ﻿using System.Text.Json;
 using Domain.Models;
+using JsonData.JsonSerializationUtils;
 
 namespace JsonData.Context;
 
 public class JsonDataContext
 {
+    private readonly IJsonHelper jsonHelper;
+
+    // todo jeg har brug for at kunne deserialize med private constructor. 
+        // https://stackoverflow.com/questions/58147552/can-i-deserialize-json-with-private-constructor-using-system-text-json
+    
     private string path = "vialabs.json";
 
     private ViaLabData? viaLabData;
+    
+
+    
 
     // todo de her skal returnere select many pga tree structurre
     public ICollection<Teacher> Teachers => ViaLabData.Teachers;
@@ -31,15 +40,16 @@ public class JsonDataContext
         }
     }
 
-    public JsonDataContext()
+    public JsonDataContext(IJsonHelper jsonHelper)
     {
+        this.jsonHelper = jsonHelper;
         if (!File.Exists(path))
         {
             ViaLabData vld = new()
             {
                 Teachers = new List<Teacher>()
                 {
-                    new Teacher("VIA\\TRMO")
+                    Teacher.Create("VIA\\TRMO").Value
                 },
                 UnApprovedTeachers = new List<Teacher>(),
             };
@@ -54,16 +64,14 @@ public class JsonDataContext
     private void LoadData()
     {
         string vldAsJson = File.ReadAllText(path);
-        viaLabData = JsonSerializer.Deserialize<ViaLabData>(vldAsJson);
+        viaLabData = jsonHelper.Deserialize<ViaLabData>(vldAsJson);
+        // int stopher = 0;
     }
 
     public void SaveChanges()
     {
         Console.WriteLine("Saving changes!");
-        string vldAsJson = JsonSerializer.Serialize(viaLabData, new JsonSerializerOptions
-        {
-            WriteIndented = true
-        });
+        string vldAsJson = jsonHelper.Serialize(viaLabData);
         File.WriteAllText(path, vldAsJson);
         viaLabData = null;
     }
@@ -74,7 +82,7 @@ public class JsonDataContext
     }
 }
 
-class ViaLabData
+public class ViaLabData
 {
     public ICollection<Teacher> Teachers { get; set; } = new List<Teacher>();
     public ICollection<Teacher> UnApprovedTeachers { get; set; } = new List<Teacher>();
